Added a "links.txt" example
This example includes three projects. The first is small and quick to analyse. The second is non-existent, so the script throws an error and continues the next link. The last link is a larger project, which has an interesting dynamic, which can be analysed in Excel sheets. 

A reminder that this file needs to be in the same, specified by the user, directory, which is set during execution of the code.